+++ src/main/java/edu/kis/vh/stacks/list/StackList.java
@@ -1,6 +1,40 @@
 package edu.kis.vh.stacks.list;
 
 public class StackList {
+	private class Node {
+
+		private final int value; // wszystkie użycia zmiennej zostały zmienione
+									// na
+									// użycia
+		// poprzez getter i setter (klasa StackList)
+		private Node prev;
+		private Node next;// tak samo jak w przypadku value
+
+		public Node(int i) {
+			value = i;
+		}
+
+		public int getValue() {
+			return value;
+		}
+
+		public Node getPrev() {
+			return prev;
+		}
+
+		public void setPrev(Node prev) {
+			this.prev = prev;
+		}
+
+		public Node getNext() {
+			return next;
+		}
+
+		public void setNext(Node next) {
+			this.next = next;
+		}
+
+	}
 
 	private static final int EMPTY_STACK = -1;
 	private Node last;
